Create Temperature
The little stuff is what's fun with this. Declaring an int without a value wasn't working right for me, so I put it at max temp, so any value is going to be closer and replace it. If there are no temps read, then it goes to zero. Then find the closest to zero. And if there's multiple values that are closest, it chooses the higher (or positive) of the two.