Create Mean Max
A work in progress, as all competition ones will be. Just grabbing the last wreck's X and Y coordinates and rushing to that. Which seems like the AI is doing too. I really need to get all of the wrecks and have it pick the closest one or overlapping ones Should be working on this soon, but love this game!