Create Mad Pod Racing
The more you try to add in some trigonometry and careful adjustments, the more you fall behind the times when you just cut thrust as you got close to the turn. Lots I need to correct here (like the adjustment needs to be smoothed out and not jerk the aim when in a different quadrant). But works well enough for now for silver.